Long 타입 테스트

+++ src/main/java/shop/mtcoding/bank/domain/account/Account.java
@@ -66,7 +66,7 @@ public class Account {
      public void checkOnwer(Long userId) {
           // String testUsername = user.getUsername();
           // System.out.println("테스트 : " + testUsername);
-          if (user.getId() != userId) { // Lazy 로등이여도 id를 조회할 때는 select 쿼리가 날라가지 않는다.
+          if (user.getId().longValue() != userId.longValue()) { // Lazy 로등이여도 id를 조회할 때는 select 쿼리가 날라가지 않는다.
                throw new CustomApiException("계좌 소유자가 아닙니다.");
           }
      }
